@@ -5,7 +5,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.18.2 <3.0.0'
+  sdk: '>=2.19.2 <3.0.0'
 
 dependencies:
   flutter:
@@ -16,21 +16,24 @@ dependencies:
 
   cupertino_icons: ^1.0.2
   flutter_zoom_drawer: ^3.0.3
-  flutter_svg: ^1.1.5
+  flutter_svg: ^2.0.2
   percent_indicator: ^4.2.2
   url_launcher: ^6.1.6
-  animate_do: ^2.1.0
+  animate_do: ^3.0.2
   shared_preferences: ^2.0.15
   get: ^4.6.5
   image_picker: ^0.8.6
   fluttertoast: ^8.1.1
   firebase_core: ^2.0.0
   firebase_auth: ^4.0.1
+  cloud_firestore: ^4.4.3
+  flutter_screenutil: ^5.6.1
+  sms_autofill: ^2.3.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.10.0
+  flutter_launcher_icons: ^0.12.0
   flutter_lints: ^2.0.0
 
 flutter_icons:
@@ -44,8 +47,8 @@ flutter:
 
   assets:
     - images/Departments/
-    - images/Login and Create Account/
-    - images/on boarding 1/
+    - images/login_and_create_account/
+    - images/on_boarding_1/
     - images/
 
   fonts:
