@@ -1,6 +1,13 @@
 class City {
-  final int id;
-  final String name;
+  late int id;
+  late String nameAr;
+  late String nameEn;
 
-  City({required this.name, required this.id});
+  City({required this.nameAr, required this.id});
+
+  City.fromMap(Map<String, dynamic> map) {
+    id = map['id'];
+    nameAr = map['name_ar'];
+    nameEn = map['name_en'];
+  }
 }
