@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:tabadul/constants/fonts.dart';
 
-import '../Widget/Button_AppBar.dart';
-import '../Widget/Buttons.dart';
+import '../Widget/custom_circle_button.dart';
+import '../Widget/custom_button.dart';
 import '../constants/colors.dart';
-import '../constants/strings.dart';
+import '../constants/routes.dart';
 
 class AnswerIsNoScreen extends StatelessWidget {
   const AnswerIsNoScreen({Key? key}) : super(key: key);
@@ -12,16 +12,16 @@ class AnswerIsNoScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: ColorsApp.colorScaffold,
+      backgroundColor:colorWhite ,
       appBar: AppBar(
         backgroundColor: Colors.transparent,
         elevation: 0,
         leadingWidth: 78,
-        leading: ButtonAppBar(
+        leading: CustomCircleButton(
           start: 40,
           paddingStart: 5,
           icon: Icons.arrow_back_ios,
-          function: () => Navigator.pop(context),
+          onPressed: () => Navigator.pop(context),
         ),
       ),
       body: ListView(
@@ -34,8 +34,8 @@ class AnswerIsNoScreen extends StatelessWidget {
             style: TextStyle(
               height: 1.8,
               fontSize: 20,
-              fontFamily: FontsApp.helveticaL,
-              color: ColorsApp.s1,
+              fontFamily: kHelveticaL,
+              color: s1,
             ),
           ),
           const SizedBox(height: 40),
@@ -63,7 +63,7 @@ class AnswerIsNoScreen extends StatelessWidget {
           Padding(
             padding:
                 const EdgeInsetsDirectional.only(top: 60, start: 5, end: 5),
-            child: Buttons(
+            child: CustomButton(
               name: 'تم',
               x: 50,
               y: 150,
@@ -85,11 +85,11 @@ class TextFieldTime extends StatelessWidget {
       keyboardType: TextInputType.datetime,
       maxLength: 1,
       textAlign: TextAlign.center,
-      cursorColor: ColorsApp.s,
+      cursorColor: primaryColor,
       style: const TextStyle(
         fontSize: 55,
         fontWeight: FontWeight.w500,
-        color: ColorsApp.s,
+        color: primaryColor,
       ),
       decoration: InputDecoration(
         constraints: const BoxConstraints(maxHeight: 80, maxWidth: 62),
@@ -99,10 +99,10 @@ class TextFieldTime extends StatelessWidget {
         hintStyle: const TextStyle(
           fontSize: 55,
           fontWeight: FontWeight.w500,
-          color: ColorsApp.s,
+          color: primaryColor,
         ),
         filled: true,
-        fillColor: ColorsApp.s4,
+        fillColor: s4,
         border: OutlineInputBorder(
           borderSide: BorderSide.none,
           borderRadius: BorderRadius.circular(10),
