@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
+import 'package:get/get.dart';
 import '../constants/fonts.dart';
-import '../constants/strings.dart';
+import '../constants/routes.dart';
 import '../constants/colors.dart';
-import '../Widget/Button_AppBar.dart';
-import '../Widget/Buttons.dart';
+import '../Widget/custom_circle_button.dart';
+import '../Widget/custom_button.dart';
 
 class WaitForPickupScreen extends StatelessWidget {
   const WaitForPickupScreen({Key? key}) : super(key: key);
@@ -12,39 +13,39 @@ class WaitForPickupScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: ColorsApp.colorScaffold,
+      backgroundColor: colorWhite,
       appBar: AppBar(
         backgroundColor: Colors.transparent,
         elevation: 0,
         centerTitle: true,
         leadingWidth: 78,
-        leading: ButtonAppBar(
+        leading: CustomCircleButton(
           start: 40,
           paddingStart: 5,
           icon: Icons.arrow_back_ios,
-          function: () => Navigator.pop(context),
+          onPressed: () => Navigator.pop(context),
         ),
       ),
       body: ListView(
         padding: const EdgeInsetsDirectional.only(start: 70, end: 70, top: 100),
         physics: const NeverScrollableScrollPhysics(),
         children: [
-          SvgPicture.asset('images/Wait.svg', fit: BoxFit.fitWidth),
+          SvgPicture.asset('images/wait.svg', fit: BoxFit.fitWidth),
           const SizedBox(height: 45),
           Text(
-            'لقد تم تحديد الوقت بنجاح\nالرجاء الانتظار حتى انتهاءالوقت\nللاستلام',
+            'the_time_has_been_set_successfully'.tr,
             textAlign: TextAlign.center,
-            style: TextStyle(
+            style: const TextStyle(
               height: 1.8,
               fontSize: 17,
-              fontFamily: FontsApp.helveticaL,
-              color: ColorsApp.s1,
+              fontFamily: kHelveticaL,
+              color: s1,
             ),
           ),
           Padding(
             padding:
                 const EdgeInsetsDirectional.only(top: 40, start: 75, end: 75),
-            child: Buttons(
+            child: CustomButton(
               name: 'إلغاء',
               x: 45,
               y: 106,
