@@ -1,54 +1,57 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 import '../constants/colors.dart';
 
-class FiledCode extends StatelessWidget {
-  const FiledCode({
+class CustomFiledCode extends StatelessWidget {
+  const CustomFiledCode({
     Key? key,
     required this.textEditingController,
     required this.focusNode,
     required this.onChanged,
+    required this.error,
   }) : super(key: key);
 
   final TextEditingController textEditingController;
   final FocusNode focusNode;
   final Function(String value) onChanged;
+  final bool error;
 
   @override
   Widget build(BuildContext context) {
-    return SizedBox(
-      width: 50,
+    return Expanded(
       child: TextFormField(
         autofocus: true,
-        enabled: true,
         controller: textEditingController,
         focusNode: focusNode,
         onChanged: onChanged,
         keyboardType: TextInputType.number,
-        cursorColor: ColorsApp.s,
+        cursorColor: primaryColor,
+        cursorRadius: const Radius.circular(20).w,
         style: TextStyle(
-          color: ColorsApp.s,
-          fontSize: 24,
-          fontWeight: FontWeight.bold,
+          color: primaryColor,
+          fontSize: 17.sp,
+          fontWeight: FontWeight.w500,
         ),
         maxLength: 1,
         textAlign: TextAlign.center,
         decoration: InputDecoration(
-          counterText: '',
-          hintText: '-',
-          hintStyle: TextStyle(
-            color: ColorsApp.s,
-            fontSize: 24,
-            fontWeight: FontWeight.bold,
+          contentPadding: EdgeInsetsDirectional.only(
+            top: 7.h,
+            bottom: 7.h,
+            start: 12.w,
+            end: 12.w,
           ),
-          contentPadding: const EdgeInsetsDirectional.only(
-            top: 15,
-            bottom: 10,
+          counterText: '',
+          hintText: '_',
+          hintStyle: TextStyle(
+            color: primaryColor,
+            fontSize: 20.sp,
           ),
           border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(10),
+            borderRadius: BorderRadius.circular(10).w,
             borderSide: BorderSide.none,
           ),
-          fillColor: ColorsApp.s4,
+          fillColor:error?kErrorColor: s4,
           filled: true,
         ),
       ),
