@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
 
-class ColorsApp {
-  static const Color s = Color(0XFF464698);
-  static const Color s1 = Color(0XFF292959);
-  static const Color s2 = Color(0XFF9090C1);
-  static const Color s3 = Color(0XFFD51818);
-  static const Color s4 = Color(0XFFF1F1F9);
-  static const Color s5 = Color(0XFFE0E0FC);
-  static const Color s6 = Color(0XFFE8E8FD);
-  static const Color s7 = Color(0XFFF7F7FE);
-  static const Color colorScaffold = Colors.white;
-}
+const Color primaryColor = Color(0XFF464698);
+const Color s1 = Color(0XFF292959);
+const Color s2 = Color(0XFF9090C1);
+const Color s4 = Color(0XFFF1F1F9);
+const Color s5 = Color(0XFFE0E0FC);
+const Color s6 = Color(0XFFE8E8FD);
+const Color s7 = Color(0XFFF7F7FE);
+const Color colorWhite = Colors.white;
+const Color kErrorColor = Color(0XFFD51818);
