@@ -1,4 +1,2 @@
-class FontsApp {
-  static String helveticaL = 'HelveticaLTArabic';
-  static String helveticaLNeeL = 'HelveticaNeeLTArabic';
-}
+const String kHelveticaL = 'HelveticaLTArabic';//bold
+const String kHelveticaLNeeL = 'HelveticaNeeLTArabic';
