@@ -5,12 +5,34 @@ const Map<String, String> ar = {
   'on_boarding2_title': 'توصيل سريع',
   'on_boarding2_subtitle':
       'نتبع سياسة توصيل آمنة تضمن وصول\nالسلعة لباب منزلك خلال 24 ساعة أو أقل',
-  'login': 'تسجيل الدخول',
-  'enter_phone_number': 'أدخل رقمك الهاتفي',
-  'do_not_have_an_account': 'ليس لديك حساب من قبل ؟ ',
-  'create_account': 'إنشاء حساب',
   'resend_code': 'إعادة إرسال الرمز',
-  'please_enter_code_correctly': 'يرجى إدخال الرمز بشكل صحيح',
   'security_code_subtitle':
       ' لقد أرسلنا لك رسالة نصية قصيرة\n تحتوي على الرمز سيتم تفعيلها تلقائيا',
+////////////////////////////////////////////////
+// Create account screen
+  'enter_mobile_number': 'أدخل رقمك الهاتفي',
+  'create_account': 'إنشاء حساب',
+  'login': 'تسجيل الدخول',
+  'login_button': 'تسجيل دخول',
+  'do_not_have_an_account': 'ليس لديك حساب من قبل ؟ ',
+  'already_have_an_account': 'هل لديك حساب من قبل ؟ ',
+//error message
+  'please_enter_code_correctly': 'يرجى إدخال الرمز بشكل صحيح',
+//
+  'the_time_has_been_set_successfully':
+      'لقد تم تحديد الوقت بنجاح الرجاء الانتظار حتى انتهاء الوقت للاستلام',
+  'cancel': 'إلغاء',
+  'notifications': 'الإشعارات',
+  'main': 'الرئيسية',
+  'profile': 'الملف الشخصي',
+  'security_code': 'رمز الأمان',
+  'add_information':'إضافة معلومات',
+  'next':'التالي',
+  'first_name':'الاسم الأول',
+  'last_name':'الاسم الأخير',
+  'city':'المدينة',
+  'area_street':'المنطقة / الشارع',
+  'the_account_has_been_created_successfully':'لقد تم إنشاء الحساب\n بنجاح',
+  'please_enter_the_number_correctly':'يرجى إدخال الرقم بشكل صحيح',
+'the_number_is_already_registered':'الرقم مسجل بالفعل',
 };
