@@ -1,13 +1,15 @@
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'shared_preference/pref_controller.dart';
 import 'package:flutter/material.dart';
 import 'languages/translation.dart';
 import 'package:get/get.dart';
+import 'utils/bindings.dart';
 import 'app_routes.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  await SharedPreferencesController().getInstance();
+  await PrefController().initializeApp();
   await Firebase.initializeApp();
   runApp(MyApp(appRoutes: AppRoutes()));
 }
@@ -18,12 +20,20 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return GetMaterialApp(
-      translations: Translation(),
-      locale: const Locale('ar'),
-      fallbackLocale: const Locale('ar'),
-      debugShowCheckedModeBanner: false,
-      onGenerateRoute: appRoutes.onGenerateRoute,
+    return ScreenUtilInit(
+      designSize: const Size(360, 690),
+      minTextAdapt: true,
+      splitScreenMode: true,
+      builder: (context, child) {
+        return GetMaterialApp(
+          translations: Translation(),
+          locale: const Locale('ar'),
+          fallbackLocale: const Locale('ar'),
+          debugShowCheckedModeBanner: false,
+          onGenerateRoute: appRoutes.onGenerateRoute,
+          initialBinding: InitialBinding(),
+        );
+      },
     );
   }
 }
