@@ -1,6 +1,7 @@
 class CodeCountry {
   final int id;
   final String name;
+  final int num;
 
-  CodeCountry({required this.name, required this.id});
+  CodeCountry({required this.name, required this.id, required this.num});
 }
